+++ src/hikari_cp/core.clj
@@ -100,12 +100,12 @@
 
 (def DatasourceClassnameConfigurationOptions
   (assoc BaseConfigurationOptions
-         :datasource-classname s/Str))
+         :datasource-class-name s/Str))
 
 ;(s/optional-key :driver-class-name)
 (def ConfigurationOptions (s/conditional
                              :datasource DatasourceConfigurationOptions
-                             :datasource-classname DatasourceClassnameConfigurationOptions
+                             :datasource-class-name DatasourceClassnameConfigurationOptions
                              :adapter AdapterConfigurationOptions
                              :jdbc-url JDBCUrlConfigurationOptions
                              :else AdapterConfigurationOptions))
@@ -144,12 +144,12 @@
         not-core-options      (apply dissoc options
                                      :username :password :pool-name :connection-test-query
                                      :configure :leak-detection-threshold :adapter :jdbc-url
-                                     :datasource-classname :driver-class-name :connection-init-sql
+                                     :datasource-class-name :driver-class-name :connection-init-sql
                                      :metric-registry :health-check-registry
                                      (keys BaseConfigurationOptions))
         {:keys [adapter
                 datasource
-                datasource-classname
+                datasource-class-name
                 auto-commit
                 configure
                 connection-test-query
@@ -185,7 +185,7 @@
       (->> (get adapters-to-datasource-class-names adapter)
            (.setDataSourceClassName config))
       (.setJdbcUrl config jdbc-url))
-    (when datasource-classname (.setDataSourceClassName config datasource-classname))
+    (when datasource-class-name (.setDataSourceClassName config datasource-class-name))
     ;; Set optional properties
     (when driver-class-name (.setDriverClassName config driver-class-name))
     (when username (.setUsername config username))